\docType{methods}
\name{as.Construct}
\alias{as.Construct}
\alias{as.Construct,data.frame-method}
\alias{as.Construct,list-method}
\title{Convert Objects to Constructs}
\usage{
\S4method{as.Construct}{list}(x, ...)

\S4method{as.Construct}{data.frame}(x, ...)
}
\arguments{
  \item{x}{an Object to convert.}

  \item{...}{optional parameters}
}
\description{
Convert Objects to Constructs

Convert Objects to Constructs
}

