\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,Construct-method}
\title{Convert Objects to data.frames}
\usage{
\S4method{as.data.frame}{Construct}(x, ...)
}
\arguments{
  \item{x}{the Construct}

  \item{...}{additional arguments to be passed to or from
  methods.}
}
\description{
Convert Objects to data.frames
}

