\docType{class}
\name{Construct}
\alias{Construct}
\alias{Construct-class}
\title{The Construct class}
\description{
This class is part of the \pkg{SemDiff} which is part of
the WMVG Project.
}
\details{
Find here the defined slots for this class.
}
\section{Slots}{
  \describe{ \item{\code{label}:}{ Object of class
  \code{"character"}, containing the name of the Construct.
  The default value will be caluclated randomly.  }
  \item{\code{rpole}:}{ Object of class \code{"character"},
  containing the name of the right pole of the construct }
  \item{\code{lpole}:}{ Object of class \code{"character"},
  containing the name of the left pole of the construct } }
}
\author{
Dipl. Kfm. Felix Lindemann
\email{felix.lindemann@hs-neu-ulm.de}

Wissenschaftlicher Mitarbeiter Kompetenzzentrum Logistik
Buro ZWEI, 17

Hochschule fur angewandte Wissenschaften Fachhochschule
Neu-Ulm | Neu-Ulm University Wileystr. 1

D-89231 Neu-Ulm

Phone +49(0)731-9762-1437 Web
\url{www.hs-neu-ulm.de/felix-lindemann/}
\url{http://felixlindemann.blogspot.de}
}

