\docType{methods}
\name{$}
\alias{$}
\alias{$,Construct-method}
\title{Extract Methods}
\usage{
\S4method{$}{Construct}(x, name)
}
\arguments{
  \item{name}{Name of Attribute.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Extract Methods
}

