\docType{methods}
\name{[}
\alias{[}
\alias{[,Construct-method}
\title{Extract Methods}
\usage{
\S4method{[}{Construct}(x, i, j, drop)
}
\arguments{
  \item{i}{Row Index}

  \item{j}{Name or Column index}

  \item{drop}{Optional value for Drop-Levels.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Extract Methods
}

