\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,Construct-method}
\title{initialize Method}
\usage{
\S4method{initialize}{Construct}(.Object, data = NULL, ...)
}
\arguments{
  \item{data}{can of type \code{\link{data.frame}} or
  \code{\link{list}}}

  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
initialize Method
}

