\docType{methods}
\name{is.Construct}
\alias{is.Construct}
\alias{is.Construct,Construct-method}
\title{Checks if an Object is as Construct}
\usage{
is.Construct(x, ...)

\S4method{is.Construct}{Construct}(x, ...)
}
\arguments{
  \item{x}{the Construct}

  \item{...}{optional Parameters (not supported)}
}
\description{
Checks if an Object is as Construct
}

