\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,Construct-method}
\title{Set Methods}
\usage{
\S4method{$}{Construct}(x, name) <- value
}
\arguments{
  \item{name}{Name of parameter to change}

  \item{value}{new Value.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Set Methods
}

